Orchestrator supports non chromium/src CLs
Orchestrator will check the project name of the current gerrit change.
If the project is not 'chromium/src' then it will pass in the head
revisions of both src and the other repo being patched (like v8).

When triggering the compilator for a non chromium/src CL, it now passes
in a 'deps_revision_override' property. Otherwise, it will default
to check out the other repo at refs/heads/main.

Also moved some of the build configuration lines to a separate function.
self.configure_build() is called in the beginning of the recipe to
keep the same behavior of the "report builders" step being up top.

Bug: 1254474, 1253997
Change-Id: Ia75aa3450e9c42b41d65dbeb8a04b99faf479f97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3195333
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>