[Flakiness] Execute the tests with (check flakiness) suffix
This CL invokes test_utils.run_tests() so that the test objects
that have had their TestOptions updated are run. It's expected
that the recipe is able to translate the TestOptions to args
for the given test type & platform, and that the underlying test
runner is able to interpret those arguments.

As a part of this CL, step names have been updated to be more
human understandable. Tests have been updated to reflect test data
closer to what we expect for the MVP on iOS.

Ensure that variants match when selecting the Test object for
the newly identified tests.

When selecting test object, we cross reference variants and dimensions,
as well as ensure that test_suite in variants == canonical name
for tests that utilize variants in testing/buildbot (these don't have
device dimensions set anywhere else aside from naming).

This CL doesn't not alter the build status yet based on the results
and has yet to be integrated into the chromium_tests trybot workflow.

Bug: 1204163
Change-Id: I2e0cc6c99c0d7b65b73447f6d02edb3316d9546e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3205881
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>