Allow iOS tests to pass --gtest_repeat and --gtest_filter arguments
As part of crrev/c/3036845, --gtest_repeat and --gtest_filter are
supported by the iOS test runner. These are expected to be used
by the flakiness prevention logic, so that a single test may be
run multiple times.

We use the TARGET_PLATFORM within chromium configs to identify
whether the tests are iOS specific.

Also, this change retains the expected file of the full.basic run
to ensure that the --gtest_repeat and --gtest_filter args
are correctly being passed within the pre_run step of (check
flakiness) suffix.

Change-Id: I067a69403e82ffa77cb64133fedd6d6b9e9f3357
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3216261
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>