Don't load the combined JSON output of swarming tests in recipe memory.
When collecting a swarming test in the chromium recipe today, we pass
in a json.output placeholder to the collect & merge script. This causes
the combined JSON output of all shards to be automatically loaded into
the recipe's memory and attached to the step_result in the "json.output"
field.

Given enough tests, and given large enough JSON blobs, that can cause
the recipe to go OOM on 2-core bots.

This CL works around that excessive RAM usage by _not_ automatically
loading the JSON blobs into the recipe's memory. Instead, if we're in
the RDB-results experiment (which should practically be 100% of all
builds at this point) we will write the combined JSON to a file we
create ourselves.

Then instead of loading the contents of the file into the recipe's mem,
we simply pass this file name down to the only user of said file
(ie: the legacy test-results upload step that FindIt still requires).

An example build that OOMs today on 2 cores when there are enough
test failures:
https://ci.chromium.org/swarming/task/568fc113540c7410?server=chromium-swarm.appspot.com

And with this patch applied, that same build (with the same amount of
failures) avoids the OOM:
https://ci.chromium.org/swarming/task/5694f08c3188d610?server=chromium-swarm.appspot.com

Bug: 1254021
Change-Id: Ice278007253006a9609d7b79f5474351289ba8b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3219634
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>