Switch code to using json API instead of json module.
The json API takes defaults to sorting keys and sets the separators so
that the same json is produced for python2 and python3. This prevents
having to go make adjustments the json code for a module when migrating
dependent modules. The sort_keys argument was preserved if it was
present before the module was marked python3 compatible.

Change-Id: I3f59ac45be68b388b13763c3c7e8b8c4a562eb68
Recipe-Nontrivial-Roll: build_limited
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3224620
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>