[V8] Show GN args in separate lines
This eases copy/pasting the GN args.

To not break ongoing builds, we keep both the old build_environment
and the new gn_args property on the builder side. In a follow up, we
remove it.

The legacy build_environment seems to be a YAGNI example. It has room
for other things than GN args, which apparently were never added in
years.

Recipe-Nontrivial-Roll: build_limited
Change-Id: Ie1a45fe27c328674c1d81de67932d78b76608cfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3236946
Reviewed-by: Liviu Rau <liviurau@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>