Change a test's RDB results from a property to a get() helper method.
The old property on the Test class exposed acces to '_rdb_results'
directly, which was problematic for experimental suites. Specifically,
experimental suites overwrite the 'suffix' used internally and don't
expose this to users. So issues often arise when an experimental
suite's RDB results are accessed without knowing the special suffix
override.

This fixes that by moving the property to a simple get() method which
takes the suffix, and overwriting that method then tweaking the suffix
arg in the experimental test class.

Bug: 1135718
Change-Id: Ib6ab32ba3f729648b4ad54196a1f88d8ae164eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3235657
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>