[flakiness] fetch results for flaky test run and update step displays
Tests run to verify flakiness aren't accurately summarized for flakiness.
The way we re-run tests for flakiness are interpreted as auto-retries
by the current implementation for ResultDB test result parsing.

This CL adds 'individual_results' as an attr to RDBPerSuiteResults,
which can store all results for a given test name. This is used by
the flakiness workflow to calculate the # of failures and determine
whether the test is flaky.

Added a TODO for updating the markdown summary of the overall result.

Bug: 1204163
Change-Id: I59e5f3530fc4444d595329b7c220a0af2b1350f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3249597
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>