publish tarball: Explicitly run clang's build.py with Python 3.
The work in bug 1261812 is switching clang's update.py to Python 3 while
build.py is still running with "python" in the shebang line.

This is being fixed in
https://chromium-review.googlesource.com/c/chromium/src/+/3253157, but
it will not fix the build of existing M97 releases that do not have it,
so explicitly invoke the script with python3 for now.

Bug: 1264763
Change-Id: Idcbd16305b078cbf62b144dc69089db9ce0ee3e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3250683
Auto-Submit: Raphael Kubo da Costa <raphael.kubo.da.costa@intel.com>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>