Add a link in led builds to the milo 'test results' tab.
led builds aren't supported in the new UI, but the new UI is capable of
rendering the 'test results' tab for them. But you need to craft the
link manually.

So this adds a step that prints that link for led builds, purely for
convenience.

Example led build with the step:
https://ci.chromium.org/swarming/task/56e0ff599cb4e910?server=chromium-swarm.appspot.com

Needs https://chromium-review.googlesource.com/c/infra/luci/recipes-py/+/3251353
before this can land.

Bug: 1264479
Change-Id: I003cda2b2dfff2371be407dfb9b357d3ff93de35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3251313
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>