chromium_tests: Add a helper util to mock RDB results for swarming tests
When switching to rdb as the source of truth for tests, the result of
a swarming test is determined by the outcome of two steps:
- the `swarming collect` step for the test
- the `rdb query` step that queries rdb for the results

We need both since a swarming shard could fail-fast without uploading
any results to RDB. In that case, RDB wouldn't indicate any failing
tests so we need to fallback on the shard's exit code.

Consequently, to mock a swarming test's results in recipe tests, you
need to mock out two different steps. That leads to excessive
boilerplate. So this helper method should cut down on that.

Bug: 1135718
Change-Id: I39cfc65c9812ef763ccac87c4d1c71dc0cc51dbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3265231
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>