Update ESLint script name
We are upgrading to ESLint 8, which requires some changes on our end. To
that extent,
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3268299
introduces a new script that is compatible with 8, but ends with `.mjs`.
Therefore, update the scripts on CQ (after the above CL landed) to start
using the new script (they are functionally equivalent).

R=liviurau@chromium.org

Bug: none
Change-Id: I8036d430ca48e0580df74c5d54212144a315b886
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3268407
Reviewed-by: Liviu Rau <liviurau@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>