Migrate all trybot_steps_with_specific_tests.py tests to RDB results.
This requires fixing shard_retry_with_patch_results() so that it can
include tests that failed in 'with patch' but passed in 'retry shards
with patch' in the return value.

This is needed because RDB only returns the tests with unexpected
results. So if a test passed as expected in the 'retry shards
with patch' phase, we wouldn't ever get its results.

Consequently, we assume any test that failed in 'with patch' but didn't
fail in 'retry shards with patch' ended up passing in the retry, and so
include it in the 'passing_tests' return tuple.

All of this is needed to maintain the step_metadata of the 'FindIt
Flakiness' step text in expectation tests. eg:
https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipe_modules/chromium_tests/tests/api/trybot_steps_with_specific_tests.py;drc=4a6f46a4bb633d9f2fbf6f51df665531c28e78f9;l=1091

Bug: 1135718
Change-Id: I4945aaffe502a6607c5a036cda416f5bb2ac1d46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3271821
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>