Add deployment to chromium_website recipe.
This CL modifies the chromium_website to deploy the built website
after it is done building.

For now, builds will be unconditionally deployed to the staging
project website by either a try job or a CI job, and if the
build is being done as part of a try job, it should post a
comment back on the CL via Tricium.

Subsequent CLs will change things so that the CI job deploys to
the live site using the production project, and
the try job deploys to preview URLs for the live site (also using
the production project).

Bug: 1267501
Change-Id: I55f9dfc5aab7c43b6a02766553888f604f835bf5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3270733
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>