Switch to prod and preview deployments for the chromium_website recipe.
Firebase Hosting supports "previews" using something called
a "channel id" when deploying content. We can use this to
stage the patchsets for a tryjob so users can preview the
results.

This CL modifies the chromium_website recipe to generate
a preview URL using a name based on the CL number and the
patchset, and posts the link to the staged website back to
the CL using the Tricium APIs.

For non-tryjobs (aka CI / post-submit jobs), this CL also
switches from using the -staging Firebase project to the
production (soon to be live) project.

Bug: 1267501
Change-Id: I6ee8f0c039d895f158659d00ebeda81fb9e2ba9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3273873
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>