[flakiness] Group multiple tests in one task when checking flakiness
This fixes suite "was already triggered" assertion failure which happens
when multiple new tests in one suite are detected, and a single task of
the same name is dispatched for each test.

This CL also extended the basic_ios test case in "examples/full.py" to
test with >1 new tests in 1 suite, and new tests in >1 suites.

Bug: 1204163
Change-Id: I210697f04fd70076707f640c84be702bccfce415
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3291889
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>