Update led_recipes_tester to use a CL on main.
led_recipes_tester is currently failing on win10 trunk jobs:
https://ci.chromium.org/ui/p/infra/builders/try/Build%20Recipes%20Tester/b8829749546567376641/overview
https://ci.chromium.org/ui/p/infra/builders/try/Build%20Recipes%20Tester/b8829746512716745233/overview

This is presumably due to the fact that it's trying jobs on a CL
that's still on master: crrev.com/c/1406154. And master is stuck on
a rev that's now 12 days old:
https://chromium.googlesource.com/chromium/src/+log/fd8a8914ca0183f0add65ae55f04e287543c7d4a

Which is leading to errors when trying to build a 12 day old revision
on newer build caches due to:
https://bugs.chromium.org/p/chromium/issues/detail?id=1272854#c11

So updating to a CL on main should hopefully fix this.

Bug: None
Change-Id: I45a91339b110b9f834cdde15c538094753922e63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3298931
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>