Switch chromium_tests swarming_isolated_script_test to using RDB results
100% of chromium builds are using RDB results. This is part of a series
of CLs to apply the same to all recipe tests based on chromium recipe
code.

This also deletes any test that exited only to provide different JSON
results, and any edge cases in the JSON contents. So since we no longer
use the JSON* we can get rid of all those tests.

*We do actually read the JSON, but only inspect the 'version' field in
it, which gates wether we run the "upload to legacy test results" step.
Unclear if that check is needed, but preserved anyway. And a test added
to explicitly cover its case.

Bug: 1135718
Change-Id: I2d51234872358852d62e7f3a40730ec999a300a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3290455
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>