Uncap the amount of results returned by rdb when querying for results.
By default, rdb will only return the first 1000 results. But we only
query for unexpected results, which cuts down on the amount of results
returned. And no suite up to this point had over 1000 unexpected
results.

However, that appears to have happened with
https://ci.chromium.org/ui/p/angle/builders/try/android-arm64-test/997/overview

A test with results (FAIL, PASS) is only tracked as (FAIL) in the
recipe because that suite has 1025 unexpected results, and the PASS
result is cutoff in that last 25.

Uncapping the limit should allow all results to be correctly
tracked in the recipe. Should any suite have so many unexpected results
that the recipe gets bogged down or crash, that's preferred over
silently ignoring every result past the 1000th.

But practically that shouldn't be much of an issue. It took 800ms
to fetch the first 1000 results in
https://ci.chromium.org/ui/p/angle/builders/try/android-arm64-test/997/overview.
That could step could easily grow to a few more seconds without issue,
which would allow 1000s more in.

Recipe-Nontrivial-Roll: build_limited
Recipe-Nontrivial-Roll: chrome_release
Bug: 1135718
Change-Id: I4fcd5a24ed070f3cc9591018442637cb29473041
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3309495
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>