Include tag in the result for precomputed try builder test history
Tags are required to determine experimental tests, so these have
been added to the BigQuery table. The changes here are to ensure
that tags are included in the resulting JSON file.

Including the tags increases the file size, and for some builders
this exceeds the extract_table -> GS bucket limit. We add a wildcard
in the result so that the data can be partitioned into x number of
JSON files.

Changes have been made to the recipe such that we can download
all the partitioned JSON files, and aggregate them into a single JSON
result for consumption. Because the files can get large, we .tar.gz
and compress for upload.

Bug: 1233132
Change-Id: Iffd65783bd049e6ac393c5568b472584e316e8b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3312363
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>