Move all legacy test_results code into boringssl-specific module.
As part of crbug.com/1135718, chromium has moved away from parsing
JSON test results to calling ResultDB's API. (This is possible because
all of chrome's tests are now uploading to ResultDB.)

Consequently, much code in the test_utils recipe module is no longer
relevant for chrome, and its only use now is the boringssl recipe.

To reflect that, and to help avoid further dependencies on locally
parsed JSON, this moves all that legacy code into a boringssl
specific recipe module.

Much of the code was copy-pasted, so this should practically be a no-op
for boringssl's builds.

Bug: 1135718
Change-Id: I0e713b14237c60157f33e19bba6f6389ba2703c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3315479
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>