Apply the RDB results recipe decision-making more broadly.
Currently, using RDB results for decision making in the chromium recipe
is controlled based on the "chromium.chromium_tests.use_rdb_results"
build experiment. Now that's it been turned on for nearly all builds,
this enables the "use_rdb_results_for_all_decisions" lever on TestSpecs
to be on by default rather than by reading the build-experiment.

This also disables that mode for all SwarmingIosTests since that's a
legacy class that:
- does not have integration with resultdb
- it would be a non-trivial amount of work to add that integration
- this class is not used on any chrome bots
- this class is only used by webrtc
- there's a bug on-file to delete this class

So it's likely safer/easier to simply disable the rdb results mode for
that class.

Bug: 1135718
Change-Id: I0a1cfb154f4cb22f385d0a87f3bc55a29fdf1edb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3318590
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>