Handle exception if latest_path does not exist
While using the experimental paths for testing the archive module on
Mac, the gsutil.download() function failed because the latest file it
was looking for did not exist in the experimental bucket. To avoid this
issue when creating a new file in other scenarios, setting the
last_version to 0.0.0.0 will allow the module to set the version to what
the build is referring to.

Bug:1260176
Change-Id: I01a784c411d8c91fef30296828d75a1002559b3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3323803
Reviewed-by: Michael Moss <mmoss@chromium.org>
Commit-Queue: Wanda Mora <morawand@google.com>