Introduce CL failure recovery mechanism
The default behavior is to mark the CL as reported by roll watcher (#rw_reported) and advertise the failure.

However alternative and potentially more complex behavior can be implemented later (screenshot updates).

Additionally we are making the search criteria a bit more flexible by:
  - allowing the subject field to be absent
  - allowing usual Gerrit terms to be enumerated in a criteria field

This was tested on the quite spamy V8 roll watcher builder.
Here we add the hashtag on the CL:
https://chrome-swarming.appspot.com/task?id=57b8f543d5a69310
And here we no longer intercept that particular CL anymore:
https://chrome-swarming.appspot.com/task?id=57b8f698164b9a10


Bug: 1237469
Change-Id: I6df12f55f34e4872e023886f83538360f68a5497
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3323256
Reviewed-by: Tamer Tas <tmrts@chromium.org>
Commit-Queue: Liviu Rau <liviurau@google.com>