Move format_step_text() from test_utils to a new recipe module.
format_step_text() is one of the few uses of test_utils in the
chromium_swarming recipe module. And if we can move it outside
test_utils, we get a little closer to removing chromium_swarming's
dependency on test_utils. And if we can fully remove that dependency,
test_utils can then safely depend on chromium_swarming. And if it
can depend on chromium_swarming safely, then we can stop passing around
'api' objects between test_utils and chromium_swarming.

In other words, we'd be able to get rid of this check:
https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipe_modules/test_utils/api.py;drc=b827d62052c857c1318150ef4eedf26e4a531113;l=741

Recipe-Manual-Change: build_limited
Bug: 1135718
Change-Id: Ieaf7b00a49d4f2b07eaa59843ec8bade7100b4e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3339538
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>