Add a diagnostic hint when builds fail to query findit for flakes.
Lately, web_test shards on mac-rel are failing 1000s of tests. This has
the side effect of making queries to findit fail with 5XX errors, eg:
See the "Failed to get known flakes" error in
https://ci.chromium.org/ui/p/chromium/builders/try/mac-rel/872757/overview

So this adds a small hint in that case to the step text as a possible
cause. We could theoretically skip calling out to findit when there are
too many failures. But the appropriate value for "too many" isn't
immediately obvious, and there doesn't appear to be much harm in just
letting the findit queries be attempted anyway.

Bug: 1281674
Change-Id: Ifd7b6529bc8b2770751982df09727266b8926279
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3352666
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>