Clarify what invalid results indicates in builds.
As part of the work in crbug.com/1135718, a bug in the chromium recipe
was fixed where a suite with a non-zero exit code but zero failing tests
would be ignored. But that's no longer the case.

Now, any suite with invalid results now likely means that the suite
did not report any failing tests but had a non-zero exit code. So this
updates the recipe failure message to indicate this.

Bug: 1280795, 1135718
Change-Id: Ie475b054be9ada83d3217054fc7eff8eb3383596
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3350761
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>