Remove some now unused RDB-migration test_utils code.
During the migration, there was two separate decision-making flows:
the old JSON-based flow and the new RDB-based flow.

The APIs that the JSON-based flow used have been migrated to RDB. So
now the two decision-making flows are both using RDB, and we can get rid
of the old experimental one.

Should be a no-op.

Bug: 1135718
Change-Id: I83bc3d93f7b5035629cf5efce80e23f9852ec647
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3378962
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>