Don't ignore NOTRUN failures that are preceded by normal FAILURE tests.
Some suites have trouble in the 'without patch' phase, and can end up
failing all their tests with NOTRUN. When those tests failed with a
normal FAILURE in the 'with patch' phase, the recipe currently considers
those two failures the same and ends up forgiving the CL of the
failures and allows it to land.

This can be problematic when the NOTRUNs are masking what otherwise
would have been PASSES, indicating an issue with the CL.

So this changes the recipe to only forgive the CL if the failing tests
have the same status (either NOTRUN or FAILURE) in both phases of the
build.

Bug: 1286397
Change-Id: I7cdb1b5ace7fd295271b8136c7305f5e56f75e4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3379757
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>