Make Android tests use SwarmingGTestTest.
The goal is to reuse Chromium's recipe code instead of relying on WebRtcIsolatedGtest which is WebRTC specific.
The 'logdog' step of android tests is remove, the logs can be found in the 'logdog' CAS output.

This is a soft revert of https://crrev.com/c/2078264.

Bug: webrtc:13556
Change-Id: Ib70a2182fb24a155ee3fe95b6562868bc93c9053
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3394586
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Jeremy Leconte <jleconte@google.com>