[benchmarking] Update recipe to recognize new builder name
Update recipe so that the mac builder will trigger the new tester mac-laptop_low_end-perf.

This is different from the CL for renaming the high end ones as we are keeping the config for the old tester, so that before the builder picks up the recipe change, the old tester will still find its config and perform recipe steps.

Bug: chromium:1279490
Change-Id: I64ad5fd425d046af45c4853beffd12e833410af4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3399585
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>