[V8] Prepare switching python resource scripts to python3
This adds convenience methods to call Python resource scripts in V8
and deprecates using the python recipe module.
In a follow up we will flip the executable to Python3 based on a luci
experiment.

The build.python recipe module will require a deprecation of using
legacy annotations in the V8 collect step first.

The scripts node_integration and auto_tag are unchanged as they will
be deprecated.

Bug: 942720, 1292003
Change-Id: I56423a7218eb6a10b94e8214df84fa9e1310077b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3423837
Reviewed-by: Liviu Rau <liviurau@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>