[V8] Collect tests without legacy annotator
This deprecates the legacy annotator in V8 and removes annotations
from the collect step. Warnings are now collected and written to
another json file in the end. As before, the warnings are written
to the step presentation as log lines, keyed by the warning title.
This preserves how the annotations were previously presented.

The legacy build.python step to collect with annotations is switched
to the convenience v8.python step to prepare the python3 migration.

The collect_v8_task.py script is very sad that it has no tests
though. We'll ensure to led test this with an older build that e.g.
had missing shards and hence triggers a warning.

Bug: 1292003
Change-Id: I73d0c9706074fef7dd8a2d183824341e0d9470e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3427204
Reviewed-by: Liviu Rau <liviurau@google.com>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>