Fix error in pin-component-crx caused by python 2 to 3 migration
The failures in the pin-component-crx builder is happening because
urlencode only accepts sequences or dictionaries. This CL will
change the dictionary of url arguments into a list of tuples so that
the order of the url arguments stays consistent when tests run under
both python 2 and python 3.

Bug:1225491
Bug:b/191630810
Change-Id: Ib502370a6a2df694410c7dc0b8e0690d71316a8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3441338
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Rakib Hasan <rmhasan@google.com>