Modify the ctbc properties builders to return the proto.
Currently the properties builder methods on the
chromium_tests_builder_config test API return a test data object from
the build method. This change modifies it so that it instead returns the
proto and the caller must pass that to the properties method of the
chromium_tests_builder_config test API.

This will allow the properties builder methods to be reused in the tests
for the builder config verifier, which doesn't need the
chromium_tests_builder_config module properties set in the test. Instead
it needs to be able to set the file content in tests to have the jsonpb
of the proto as the content of properties files.

Bug: 1288604
Change-Id: Icb5382b944a893ac392638fd87fd91d350e3a54b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3441337
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>