Attach the chromium_tests API object to every steps.Test class
The steps.py classes are mainly constructs of the chromium_tests recipe
module. But since they're defined in a file/class outside the immediate
ChromiumTestsApi class, they don't have easy access to a RecipeApi
object, which is needed for accessing any other module's utilities (eg:
api.file.read() or api.json.output()).

Today, this is solved by having clients pass in a "caller_api" to nearly
every method of the class. This has a few drawbacks:
- It makes the API of the steps.Test classes clunky. Any non-trivial
  method has the `caller_api` repeated in its arg list.
- Other client recipe modules often naturally pass their own recipe API
  object. (eg: https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipe_modules/ios/api.py;drc=282f4e63e49bc7164c0be0e8cdb163cc38e04268;l=1041)
  This requires that they mirror chromium_tests' recipe dependencies in
  their own dependency list, which can lead to weird checks like:
https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipe_modules/test_utils/api.py;drc=5ba04db5701ae0dc476bae6ac44a07974e2a3db9;l=609

By requiring that the chromium_tests recipe API object be passed on
test creation, we can store the API object and use it for all future API
accesses later on in the test.

Ideally, most creation of these tests use the factories in
generators.py, which would abstract away the needed constructor arg.
But there are some clients (eg: webrtc) that construct them directly
and so need updating as well.

See crrev.com/c/3345551 for what this would make the recipe code look
like in the end.

Bug: 1135718
Change-Id: I1526b6d21d99ce28476dc9028c89b85a08db77d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3437769
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>