Fix conversion of RTS condition values.
Whereas in JSONPB enums are repesented as the string with the name of
the enum value, in proto objects they are the integer value of the enum
value. To match our use of strings for the condition value in python,
conversion is necessary.

Bug: 1248219
Change-Id: I99b5c5e10a176c209c63f356001ca648d4c8d84a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3457649
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>