Report implicitly mirrored testers in mirrored_builders property.
A try builder can indicate that it mirrors all of the testers that are
triggered by a builder by setting include_all_triggered_testers in the
try spec. In this case, the ID for the testers will not necessarily
appear in the try spec itself but the triggered testers will be treated
as mirrors by the try builder.

Change-Id: I5fa37c23db2acc02d7c56c713ea795ad5fc43917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3470272
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@google.com>