Remove mmoss@chromium.org from recipes/recipe_modules/chromiumdash/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I00e69c798f80788080d06dcb1497b16b1d9e3228
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3501735
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Commit-Queue: Sven Zheng <svenzheng@chromium.org>