Remove weili@chromium.org from recipes/recipes/pdfium.expected/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ic2a793c1c52bd39ea7e3fa37c3bcf7d7e042f686
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3501008
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>