Orchestrator handles compilator_watcher cancel status
Previously when the CQ would cancel builds, the orchestrator build
would be set to CANCELED and overwrite any step failures that occurred
in the recipe during the cancelation. Now with a new cancel process, the
CANCELED status set by the CQ cancel request is overwritten by the
recipe's step failures. So instead of the orchestrator displaying a
CANCELED status, it is showing an INFRA_FAILURE when it is canceled
while the compilator_watcher is still running (since the
compilator_watcher is currently returning an INFRA_FAILURE status when
its context is canceled).
See thread: https://chat.google.com/room/AAAAXGgrwSo/HXUpu1WtcTM

The compilator_watcher sub_build status will be set as
CANCELED instead of INFRA_FAILURE (see CL link) when the orchestrator
build is canceled while the compilator_watcher is running.
[CL]: https://chromium-review.googlesource.com/c/infra/infra/+/3513884

The orchestrator will see that the sub_build has a CANCELED status and
check if the orchestrator build itself is also in global shutdown mode.
If so, the orchestrator will set its build status to CANCELED.

There should still never be a reason for the compilator to be canceled
without the orchestrator being canceled too (unless someone manually
goes in and kills the compilator's swarming task), so in that case the
orchestrator will raise an INFRA_FAILURE.

Bug: 1304475
Change-Id: Ia04cd656c131d03c88db96ba8e75fbc21ed103cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3514472
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Stephanie Kim <kimstephanie@google.com>