Remove sarahmashay@google.com from recipes/recipes/tint.expected/OWNERS
This is an automatically generated changelist based on group memberships and affiliations, with some transition delay.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

To report an issue, file a bug in the Infra>Security component.

Change-Id: I6d1ed9a1f2dc1f11daa6f9cb4e2e1c1a5afa5e10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3522546
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Ryan Harrison <rharrison@google.com>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>