Don't set buildbucket build in reclient.properties test API method.
Methods named properties on a test API are conventionally used only for
setting the properties for that module. Setting buildbucket data in such
a method is surprising and makes it difficult to use the properties
method when trying to create a test case that would emulate a different
build.

Change-Id: I0f2f2c9485c07caf25af6101f2e5c6507118e218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3523527
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>