Add chromium/builder_config_migration recipe.
The chromium/builder_config_migration provides functionality for
assisting in migrating the builder config src-side. Currently, the
recipe provides the ability to compute the groupings of builders that
need to be migrated at the same time and output this information to a
file. The grouping information will be used to add a presubmit that
prevents the addition of config that creates a new grouping while
allowing addition of configs that augment existing groupings.

An additional operation will be added in the future that will output the
edits that need to be made to the starlark code to migrate the config
src-side for a given builder and all related builders.

Bug: 868153
Change-Id: Id729f8d890f950ca13596a0ad64be981830196f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3523525
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>