Add the enable_reclient gclient config if reclient is configured.
This will ensure that the reclient package is downloaded if the builder
sets an reclient instance in its properties. This makes migrating to
reclient entirely driven by the builder properties for recipes that use
chromium_tests.configure_build so that changes only need to be made in
src for builders using those recipes.

Change-Id: Idcd6d2eb0686e70ad17cf6193fde106b1ebd458c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3523745
Reviewed-by: Bozydar Szadkowski <bozydar@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>