Update ninja_wrapper_test to not leave files in working directory.
Currently a file named target3 is created in the current directory when
ninja_wrapper_test is run because calling ninja_wrapper.main with the -o
flag set causes it to write the results json to a file. This changes it
to write the file to a temp directory instead.

Change-Id: I82e67ce9a53cffa3d3dfa548fa49a5361d7e9682
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/tools/build/+/3533613
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Gregory Guterman <guterman@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>